+++ Gopkg.toml
@@ -1,0 +1,19 @@
+[[constraint]]
+  name = "github.com/pkg/errors"
+  version = "0.8.0"
+
+[[constraint]]
+  branch = "master"
+  name = "golang.org/x/net"
+
+[[constraint]]
+  branch = "v1"
+  name = "gopkg.in/check.v1"
+
+[[constraint]]
+  name = "gopkg.in/fatih/pool.v2"
+  version = "2.0.0"
+
+[prune]
+  go-tests = true
+  unused-packages = true
